requirements.txt
alembic==1.15.2; python_version >= "3.9"
annotated-types==0.7.0; python_version >= "3.8"
anyio==4.9.0; python_version >= "3.9"
async-timeout==5.0.1; python_version < "3.11.0" and python_full_version >= "3.8.0" and python_version >= "3.8"
asyncpg==0.30.0; python_full_version >= "3.8.0"
certifi==2025.1.31; python_version >= "3.6"
charset-normalizer==3.4.1; python_version >= "3.7"
click==8.1.8; python_version >= "3.7"
colorama==0.4.6; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.7.0")
dnspython==2.7.0; python_version >= "3.9"
email-validator==2.2.0; python_version >= "3.8"
exceptiongroup==1.2.2; python_version < "3.11" and python_version >= "3.9"
fastapi-cli==0.0.7; python_version >= "3.8"
fastapi==0.115.12; python_version >= "3.8"
greenlet==3.1.1; python_version >= "3.7"
h11==0.14.0; python_version >= "3.7"
httpcore==1.0.7; python_version >= "3.8"
httptools==0.6.4; python_full_version >= "3.8.0"
httpx==0.28.1; python_version >= "3.8"
idna==3.10; python_version >= "3.6"
iniconfig==2.1.0; python_version >= "3.8"
jinja2==3.1.6; python_version >= "3.7"
mako==1.3.9; python_version >= "3.8"
markdown-it-py==3.0.0; python_version >= "3.8"
markupsafe==3.0.2; python_version >= "3.9"
mdurl==0.1.2; python_version >= "3.7"
packaging==24.2; python_version >= "3.8"
pluggy==1.5.0; python_version >= "3.8"
psycopg2==2.9.10; python_version >= "3.8"
pydantic-core==2.33.0; python_version >= "3.9"
pydantic==2.11.0; python_version >= "3.9"
pygments==2.19.1; python_version >= "3.8"
pytest-asyncio==0.26.0; python_version >= "3.9"
pytest==8.3.5; python_version >= "3.8"
python-dotenv==1.1.0; python_version >= "3.9"
python-multipart==0.0.20; python_version >= "3.8"
pyyaml==6.0.2; python_version >= "3.8"
requests==2.32.3; python_version >= "3.8"
rich-toolkit==0.14.0; python_version >= "3.8"
rich==13.9.4; python_full_version >= "3.8.0"
shellingham==1.5.4; python_version >= "3.7"
sniffio==1.3.1; python_version >= "3.7"
sqlalchemy==2.0.40; python_version >= "3.7"
starlette==0.46.1; python_version >= "3.9"
tomli==2.2.1; python_version < "3.11" and python_version >= "3.9"
typer==0.15.2; python_version >= "3.7"
typing-extensions==4.13.0; python_version >= "3.8"
typing-inspection==0.4.0; python_version >= "3.9"
urllib3==2.3.0; python_version >= "3.9"
uvicorn==0.34.0; python_version >= "3.9"
uvloop==0.21.0; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.9" and python_full_version >= "3.8.0"
watchfiles==1.0.4; python_version >= "3.9"
websockets==15.0.1; python_version >= "3.9"
